Implement automated editing process

+++ src/api/drafts.rs
@@ -8,16 +8,27 @@ use actix_web::{
     Responder,
     http::Method,
 };
+use failure::Fail;
 use futures::{Future, future};
 use serde::{Deserialize, Serialize};
 use uuid::Uuid;
 
 use crate::{
+    ApiError,
+    db::types::SlotPermission,
     models::{
         File,
-        draft::{Draft, PublicData as DraftData},
+        user::{User, Fields, PublicData as UserData},
+        editing::{Slot, VersionData, SlotData},
+        draft::{
+            Draft,
+            PublicData as DraftData,
+            AdvanceResult as DraftAdvanceResult,
+            FindDraftError,
+        },
+        module::PublicData as ModuleData,
     },
-    processing::ProcessDocument,
+    permissions::ManageProcess,
 };
 use super::{
     Error,
@@ -25,6 +36,7 @@ use super::{
     RouterExt,
     State,
     session::Session,
+    util::FormOrJson,
 };
 
 /// Configure routes.
@@ -34,9 +46,8 @@ pub fn routes(app: App<State>) -> App<State> {
         .resource("/drafts/{id}", |r| {
             r.get().api_with(get_draft);
             r.put().api_with(update_draft);
-            r.delete().api_with(delete_draft);
         })
-        .api_route("/drafts/{id}/save", Method::POST, save_draft)
+        .api_route("/drafts/{id}/advance", Method::POST, advance_draft)
         .resource("/drafts/{id}/comments", |r| {
             r.get().f(list_comments);
             r.post().f(add_comment);
@@ -48,6 +59,8 @@ pub fn routes(app: App<State>) -> App<State> {
             r.delete().api_with(delete_file);
         })
         .api_route("/drafts/{id}/books", Method::GET, list_containing_books)
+        .api_route("/drafts/{id}/process", Method::GET, get_process_details)
+        .api_route("/drafts/{id}/process/slots/{slot}", Method::PUT, assign_slot)
 }
 
 type Result<T, E=Error> = std::result::Result<T, E>;
@@ -65,7 +78,11 @@ pub fn list_drafts(
 ) -> Result<Json<Vec<DraftData>>> {
     let db = state.db.get()?;
     let drafts = Draft::all_of(&*db, session.user)?;
-    Ok(Json(drafts.into_iter().map(|d| d.get_public()).collect()))
+
+    drafts.into_iter().map(|d| d.get_public(&*db, session.user_id()))
+        .collect::<Result<Vec<_>, _>>()
+        .map(Json)
+        .map_err(Into::into)
 }
 
 /// Get a draft by ID.
@@ -81,9 +98,9 @@ pub fn get_draft(
     id: Path<Uuid>,
 ) -> Result<Json<DraftData>> {
     let db = state.db.get()?;
-    let draft = Draft::by_id(&*db, *id, session.user)?;
+    let draft = Draft::by_id_and_user(&*db, *id, session.user)?;
 
-    Ok(Json(draft.get_public()))
+    draft.get_public(&*db, session.user_id()).map(Json).map_err(Into::into)
 }
 
 #[derive(Debug, Deserialize)]
@@ -105,57 +122,64 @@ pub fn update_draft(
     update: Json<DraftUpdate>,
 ) -> Result<Json<DraftData>> {
     let db = state.db.get()?;
-    let mut draft = Draft::by_id(&*db, *id, session.user)?;
+    let mut draft = Draft::by_id_and_user(&*db, *id, session.user)?;
+
+    if !draft.check_permission(&*db, session.user_id(), SlotPermission::Edit)? {
+        return Err(InsufficientSlotPermission(SlotPermission::Edit).into());
+    }
 
     draft.set_title(&*db, &update.title)?;
 
-    Ok(Json(draft.get_public()))
+    draft.get_public(&*db, session.user_id()).map(Json).map_err(Into::into)
 }
 
-/// Delete a draft
+#[derive(Deserialize)]
+pub struct AdvanceData {
+    target: i32,
+    slot: i32,
+}
+
+/// Advance a draft to the next editing step.
 ///
 /// ## Method
 ///
 /// ```text
-/// DELTE /drafts/:id
+/// POST /drafts/:id/advance
 /// ```
-pub fn delete_draft(
+pub fn advance_draft(
     state: actix_web::State<State>,
     session: Session,
     id: Path<Uuid>,
-) -> Result<HttpResponse> {
+    form: FormOrJson<AdvanceData>,
+) -> Result<Json<AdvanceResult>> {
     let db = state.db.get()?;
-    let draft = Draft::by_id(&*db, *id, session.user)?;
+    let form = form.into_inner();
+    let draft = Draft::by_id_and_user(&*db, *id, session.user)?;
 
-    draft.delete(&*db)?;
-
-    Ok(HttpResponse::Ok().finish())
+    draft.advance(&*db, session.user_id(), form.slot, form.target)
+        .and_then(|r| Ok(match r {
+            DraftAdvanceResult::Advanced(draft) => AdvanceResult::Advanced {
+                draft: draft.get_public_small(),
+            },
+            DraftAdvanceResult::Finished(module) => AdvanceResult::Finished {
+                module: module.get_public(&*db)?,
+            },
+        }))
+        .map(Json)
+        .map_err(Into::into)
 }
 
-/// Save a draft.
-///
-/// ## Method
-///
-/// ```text
-/// POST /drafts/:id/save
-/// ```
-pub fn save_draft(
-    state: actix_web::State<State>,
-    session: Session,
-    id: Path<Uuid>,
-) -> Result<HttpResponse> {
-    let db = state.db.get()?;
-    let draft = Draft::by_id(&*db, *id, session.user)?;
-
-    if let Err(err) = state.xref_processor.try_send(ProcessDocument {
-        document: (**draft).clone(),
-    }) {
-        error!("Could not send document {} for processing: {}", draft.id, err);
+#[derive(Serialize)]
+#[serde(tag = "code")]
+pub enum AdvanceResult {
+    #[serde(rename = "draft:process:advanced")]
+    Advanced {
+        draft: DraftData,
+    },
+    #[serde(rename = "draft:process:finished")]
+    Finished {
+        module: ModuleData,
     }
-
-    draft.save(&*db)?;
-
-    Ok(HttpResponse::Ok().finish())
 }
 
 /// Get comments on a draft.
@@ -199,7 +223,7 @@ pub fn list_files(
     id: Path<Uuid>,
 ) -> Result<Json<Vec<FileInfo>>> {
     let db = state.db.get()?;
-    let draft = Draft::by_id(&*db, *id, session.user)?;
+    let draft = Draft::by_id_and_user(&*db, *id, session.user)?;
 
     let files = draft.get_files(&*db)?
         .into_iter()
@@ -226,7 +250,7 @@ pub fn get_file(
 ) -> Result<impl Responder> {
     let (id, name) = path.into_inner();
     let db = state.db.get()?;
-    let draft = Draft::by_id(&*db, id, session.user)?;
+    let draft = Draft::by_id_and_user(&*db, id, session.user)?;
 
     Ok(draft.get_file(&*db, &name)?
         .stream(&state.config))
@@ -253,10 +277,17 @@ pub fn update_file(
         Err(err) => return Box::new(future::err(err.into())),
     };
 
-    let draft = match Draft::by_id(&*db, id, session.user) {
+    let draft = match Draft::by_id_and_user(&*db, id, session.user) {
         Ok(draft) => draft,
         Err(err) => return Box::new(future::err(err.into())),
     };
+
+    match draft.check_permission(&*db, session.user_id(), SlotPermission::Edit) {
+        Ok(true) => (),
+        Ok(false) => return Box::new(future::err(
+            InsufficientSlotPermission(SlotPermission::Edit).into())),
+        Err(err) => return Box::new(future::err(err.into())),
+    }
 
     Box::new(File::from_stream::<_, _, Error>(
             state.db.clone(),
@@ -284,7 +315,11 @@ pub fn delete_file(
 ) -> Result<HttpResponse> {
     let (id, name) = path.into_inner();
     let db = state.db.get()?;
-    let draft = Draft::by_id(&*db, id, session.user)?;
+    let draft = Draft::by_id_and_user(&*db, id, session.user)?;
+
+    if !draft.check_permission(&*db, session.user_id(), SlotPermission::Edit)? {
+        return Err(InsufficientSlotPermission(SlotPermission::Edit).into());
+    }
 
     draft.delete_file(&*db, &name)?;
 
@@ -304,9 +339,87 @@ pub fn list_containing_books(
     id: Path<Uuid>,
 ) -> Result<Json<Vec<Uuid>>> {
     let db = state.db.get()?;
-    let draft = Draft::by_id(&*db, *id, session.user)?;
+    let draft = Draft::by_id_and_user(&*db, *id, session.user)?;
 
     draft.get_books(&*db)
         .map(Json)
         .map_err(Into::into)
 }
+
+#[derive(Serialize)]
+pub struct SlotSeating {
+    #[serde(flatten)]
+    slot: SlotData,
+    user: Option<UserData>,
+}
+
+#[derive(Serialize)]
+pub struct ProcessDetails {
+    #[serde(flatten)]
+    process: VersionData,
+    slots: Vec<SlotSeating>,
+}
+
+/// Get details of the process this draft follows.
+///
+/// ## Method
+///
+/// ```text
+/// GET /drafts/:id/process
+/// ```
+pub fn get_process_details(
+    state: actix_web::State<State>,
+    session: Session<ManageProcess>,
+    id: Path<Uuid>,
+) -> Result<Json<ProcessDetails>> {
+    let db = state.db.get()?;
+    let draft = Draft::by_id(&*db, *id)?;
+    let process = draft.get_process(&*db)?;
+
+    if !draft.check_access(&*db, &session.user(&*db)?)? {
+        return Err(FindDraftError::NotFound.into());
+    }
+
+    let slots = process.get_slots(&*db)?
+        .into_iter()
+        .map(|slot| Ok(SlotSeating {
+            slot: slot.get_public(),
+            user: slot.get_occupant(&*db, draft.module_id())?
+                .map(|user| user.get_public(Fields::empty())),
+        }))
+        .collect::<Result<Vec<_>>>()?;
+
+    Ok(Json(ProcessDetails {
+        process: process.get_public(),
+        slots,
+    }))
+}
+
+/// Assign a specific user to a slot.
+///
+/// ## Method
+///
+/// ```text
+/// PUT /drafts/:id/process/slots/:slot
+/// ```
+pub fn assign_slot(
+    state: actix_web::State<State>,
+    _session: Session<ManageProcess>,
+    path: Path<(Uuid, i32)>,
+    user: Json<i32>,
+) -> Result<HttpResponse> {
+    let (draft, slot) = path.into_inner();
+    let db = state.db.get()?;
+    let draft = Draft::by_id(&*db, draft)?;
+    let slot = Slot::by_id(&*db, slot)?;
+    let user = User::by_id(&*db, *user)?;
+
+    slot.fill_with(&*db, draft.module_id(), &user)?;
+
+    Ok(HttpResponse::Ok().finish())
+}
+
+#[derive(ApiError, Debug, Fail)]
+#[fail(display = "Missing required slot permission '{}'", _0)]
+#[api(code = "draft:process:insufficient-permission", code = "FORBIDDEN")]
+struct InsufficientSlotPermission(SlotPermission);
